chore: upgraded dependencies
