+++ client/@types/global.d.ts
@@ -1,0 +1,5 @@
+export declare global {
+  interface Window {
+    stardust: { namedRoutes: Record<string, string> };
+  }
+}
