+++ client/index.ts
@@ -1,0 +1,1 @@
+export { initRoutes, stardust } from './Stardust';
