+++ package.json
@@ -24,7 +24,7 @@
     "clean": "rimraf build",
     "build": "cross-env npm run clean && npm run build:node && npm run build:client",
     "build:node": "tsc",
-    "build:client": "esbuild ./build/client/stardust.js --outfile=build/client/stardust.js --allow-overwrite --bundle --platform=node --target=chrome91,edge90,firefox90,safari13",
+    "build:client": "esbuild ./build/client/index.js --outfile=build/client/index.js --allow-overwrite --bundle --platform=node --target=chrome91,edge90,firefox90,safari13",
     "watch": "cross-env npm run clean && tsc -w",
     "test": "nyc node japaFile.js"
   },
@@ -32,17 +32,17 @@
     "@adonisjs/core": ">=5"
   },
   "devDependencies": {
-    "@adonisjs/core": "^5.3.2",
+    "@adonisjs/core": "^5.3.4",
     "@adonisjs/mrm-preset": "^4.1.2",
     "@adonisjs/require-ts": "^2.0.8",
-    "@adonisjs/view": "^6.1.0",
+    "@adonisjs/view": "^6.1.1",
     "@commitlint/cli": "^13.1.0",
     "@commitlint/config-conventional": "^13.1.0",
     "@commitlint/prompt-cli": "^13.1.0",
     "adonis-preset-ts": "^2.1.0",
     "copyfiles": "^2.4.1",
     "cross-env": "^7.0.3",
-    "esbuild": "^0.12.25",
+    "esbuild": "^0.12.28",
     "eslint": "^7.32.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-adonis": "^1.3.3",
@@ -51,9 +51,9 @@
     "japa": "^3.1.1",
     "lint-staged": "^11.1.2",
     "nyc": "^15.1.0",
-    "prettier": "^2.3.2",
+    "prettier": "^2.4.0",
     "rimraf": "^3.0.2",
     "semantic-release": "^17.4.7",
-    "typescript": "^4.4.2"
+    "typescript": "^4.4.3"
   }
 }
